project_PTDL_demo2/postgres_to_ssms/requirements.txt
pandas
psycopg2-binary
sqlalchemy
pyodbc
